chore(canary): Avoid `workspace:*` in published package.json files (#10532)
**Problem**
It appears leaving these `workspace:*` values in for peer dependencies
can result in:
```
➤ YN0001: │ TypeError: Cannot read properties of null (reading 'set')
```
This was happening to me today. We can simply replace these with the
canary version we are publishing.

We already avoid `workspace:*` in our stable releases.

**Changes**
Updates the existing shell script to also replace all occurrences of
`workspace:*` with the canary version.