fix(vite): Rename `serverAuthContext` to `serverAuthState` (#10653)
**Problem**
https://github.com/redwoodjs/redwood/pull/10643 did some renaming but
looks like there was just one more `serverAuthContext` that needed
switched over to `serverAuthState`. I noticed this just because the CI
was failing on renovate PRs.

**Note**
I'll merge this now and we can discuss retroactively if I misinterpreted
anything since I haven't been in the middleware space as much. It's
awesome work can't wait for all this to get out into stable for people
to enjoy!