fix(dbAuthMw): Update and fix logic related to dbAuth "verbs" and decryptionErrors (#10668)
I found some issues with our dbAuth MW as I was implementing a new
feature. I think the problems are specific to when you use a
`cookieName` which is why we didn't

This PR does the following:
- updates the dbauth mw to correctly handle the cookieName option (it
should always have been optional)
- throws an error when the `dbAuthSession` returns an empty decoded
token so that it clears the authState
- we had a check for only "POST" requests to be passed to the
dbAuthHandler. This was incorrect because some of the dbAuth "verbs" or
actions - like `webAuthnRegOptions` - uses a GET request.

As a result, the tests started showing failures, so I:
- added a mock for `dbAuthSession`, so we can check both happy path and
unhappy paths for session decryption
- updated the tests where relevant