Updates seeds script to be generic, no more errors throw on first run (#10739)
If you edit the `schema.prisma` file and change the table name before
running `yarn rw prisma migrate dev` the first time, you'll get a scary
error at the end of your run:

![Image](https://github.com/orgs/redwoodjs/projects/18/assets/300/e0af73f0-3fa9-4b49-bfc9-b54fe9d8a868)

The error is caught, and doesn't stop the migration from running, but
looks nasty. Everyone will see this when running through the tutorial as
the first thing we do is change the `schema.prisma` file and then create
a migration.

The seed should not contain any actual seed data, just comments about
how to use it when the time comes. I've updated the template to be much
simpler to start, and includes a link to a new doc I created that goes
into more detail about how to use seeds, best practices, etc. The first
time you migrate the console message will point you to the file, and the
file will point you to the docs.

---------

Co-authored-by: Josh GM Walker <56300765+Josh-Walker-GM@users.noreply.github.com>