fix: smol-toml imports (#10764)
**Problem**
There is no default export from the `smol-toml` package and so usage
such as that shown below does not quite work.
```ts
import toml from 'smol-toml'
toml.parse(...)
```
Instead the easiest change is to adopt:
```ts
import * as toml from 'smol-toml'
toml.parse(...)
```
which was already being used within a test file.

**Changes**
1. Uses namespace imports for the smol toml package where that package
is used.

**Update**
It is somewhat confusing that the `import toml from 'smol-toml'` does
not appear to be working given that the documentation for the package
says that you should be able to do things that way. I'll see if this
change fixes the bug that I believe is happening. We can easily revert
or update the usage here when we understand better.