fix(cli): Add check for excessively long route (#10830)
**Problem**
Code scanning is alerting to the "Polynomial regular expression used on
uncontrolled data" issue associated with that route processing.

It isn't really that big a problem. This is CLI code which operates on
your own project. It is not really any kinda of attack vector which
would let someone harm someones project/code/product.

**Changes**
1. Adds a length check to the route string. We assume all well behaved
route definitions are less than 2000 characters - an absurdly generous
limit.