chore(fixtures): change redwood versions in fixtures (#10832)
**Problem**
Security alerts are firing on the fixture projects. These projects list
very old versions of redwoodjs and as such are triggering warnings for
old transient dependencies.

**Changes**
1. Switch to the "latest" tag in these fixtures.

**Notes**
It would be ideal if these fixtures were kept up to date like our main
test project fixture. The fixtures changed here are used only for their
structure and content - they don't currently undergo any yarn installs.

In the future I would suggest we remove these fixtures entirely if it is
possible and convenient to do so.