chore(test-project): switch to tarsync (#10962)
The ts-to-js part of generating a javascript test project was broken. I
think this is because it was not properly synced. It was complaining
about the typescript prettier plugin. If I recall correctly the issue
was because we upgraded to v5 and then you have to use esm which is why
we now have a simple string rather than a require statement in the
config file.