chore(publint): fix concurrently usage (#10969)
CI is flaky on the new `check:package`. I think this is because of how
the command is written with concurrently. Currently it is written as
`"concurrently npm:check:attw yarn publint"` which is running 3 commands
as `yarn publint` is not quoted and so isn't being treated as one single
command. I've updated the format to be: `"concurrently npm:check:attw
yarn:publint"` which works locally and I hope will fix CI.