feat(trusted-docs): Allows useRedwoodTrustedDocuments to set more custom UsePersistedOperationsOptions (#10894)
Allows useRedwoodTrustedDocuments to set more custom
UsePersistedOperationsOptions

Allows the useRedwoodTrustedDocuments plugin to define:

```ts
 /**
   * Whether to allow execution of arbitrary GraphQL operations aside from persisted operations.
   */
  allowArbitraryOperations?: boolean | AllowArbitraryOperationsHandler;
  /**
   * The path to the persisted operation id
   */
  extractPersistedOperationId?: ExtractPersistedOperationId;

  /**
   * Whether to skip validation of the persisted operation
   */
  skipDocumentValidation?: boolean;

  /**
   * Custom errors to be thrown
   */
  customErrors?: CustomPersistedQueryErrors;
```

This can let you override to allow certain ops or skip validation etc:

> If you validate your persisted operations while building your store,
we recommend to skip the validation on the server. So this will reduce
the work done by the server and the latency of the requests.

The allow authenticated request is still considered, but
`allowArbitraryOperations` can override.

Omitted `getPersistedOperation` as this extracts hash from store.