fix(realtime/trusted-docs): Supports GraphQL subscriptions and trusted documents (#10893)
Fixes: https://github.com/redwoodjs/redwood/issues/10892

This PR updates the SSELink to check if there is a trusted document hash
in the request. If there is, then don't also include the query.

The persisted operations plugin checks if the params has a query. If it
does then raises an error that only persisted operations are allowed.

Subscriptions failed this test because the request had both the hash and
a query. And since there were query details, the request was blocked.

Now, if there is a hash, the SSELink won't add back the query and thus
it succeeds.

---------

Co-authored-by: Josh GM Walker <56300765+Josh-Walker-GM@users.noreply.github.com>