fix(deps): update dependency ts-morph to v22 (#11041)
This PR updates the version of `ts-morph` to v22. 

The various release notes:
https://github.com/dsherret/ts-morph/releases?page=1

You'll see that most of the breaking changes seem to be related to
changes in TS version. The other handful of breaking changes are
difficult to evaluate. I have done my best to determine if any of them
effect us and haven't noticed any. I will have to rely on CI to confirm.