Update GraphQL docs for Docusaurus graphql documentation generator plugin (#11102)
Hey :wave: 

I am the maintainer of `@graphql-markdown/docusaurus`, formally
`@edno/docusaurus2-graphql-doc-generator`.

Thank you for referencing the plugin in your docs, that's just awesome
:heart:

I've noticed that the doc still points to the old name of the package
that is now deprecated. So, I've brushed up the page with updated
information. I hope that's all fine by you.