Disable unused imports/ vars/ params in TypeScript compiler, opt for ESLint instead (#11125)
When I'm coding against the framework I'm often leaving the files in a
"dirty state" in order to validate an idea. They're not logical errors,
but merely unused imports, variables and parameters. This often leaves
me having to comment out large swathes of code to test my idea and slows
me down!

I polled the core contributors and the majority agreed that this would
be helpful, so in this PR I disable the unused variable checking in
TypeScript and make ESLint check for it instead.