fix: Comment out unused import line in seed script template (#11171)
We changed the seed script to no include a bunch of actual database
steps but still kept it in as comments to show the user an example of
how to do it.

I think instead of adding a skip about the unused import we could just
comment it out too. Thoughts @cannikin?