Docs: update GitHub workflow file for running tests (#11173)
I tried following the guide and added the workflow to my repo but then
the action failed with the following error:
```
error This project's package.json defines "packageManager": "yarn@4.1.0". However the current global version of Yarn is 1.22.22.

Presence of the "packageManager" field indicates that the project is meant to be used with Corepack, a tool included by default with all official Node.js distributions starting from 16.9 and 14.19.
Corepack must currently be enabled by running corepack enable in your terminal. For more information, check out https://yarnpkg.com/corepack.
Error: Process completed with exit code 1.
```
I enabled Corepack and the workflow run correctly.

Added the extra step in the docs for the people that might incur in the
same issue.
Also updated the `actions/checkout` and `actions/setup-node` to v4.