fix: Update default tsconfig options (target, module and moduleResolution) (#11170)
A window into my thinking here:

`target`
Set to `ES2023` for both API and web sides. I switched from `ESNext`
just so that we are on a pinned version.

`moduleResolution`
For the API side this is `Node16`. This was changed from `Node` to
reflect the true module resolution behaviour of the modern version of
node that your code will be built for. For example taking into account
`exports` in the package json file.
For the web side this is `Bundler`. This reflects the fact that we will
be bundling your code with vite/esbuild/whatever and that TS should not
necessarily enforce rules which will not be required - like nodes
relative path extensions with import statements. This is the option that
I'm least happy with as I have a slight fear that choosing `Bundler`
when the transpiled code will ultimately run in a nodejs environment for
things like SSR might result in issues.

`module`
For the API side this is `Node16` as constrained by my choice of
`moduleResolution`.
For the web side this is `ES2022`. This was simply the most modern
version corresponding to the `ES2023` module option without going to
`ESNext` - again just because I wanted to pin these.

Concerns:
1. When you are outputting esm and not cjs you'll have to use extensions
(.js) on relative imports on the API side. Stable redwood still
transpiles to cjs for the api side.