chore(lint): first refactoring towards enabling 'require-await' rule (#11279)
We will look to enable the `typescript-eslint/require-await` rule in the
future. Currently we have many issues flagged when we do. This is the
first work to refactor to correct issues related to that rule.

For now I'm focusing on tests so that we can be sure none of this is
user facing. The changes here are mostly just removing unneeded `async`
keywords.