chore: delete crowdin config file (#11291)
I don't really have any context around this file. I wasn't with the
project when it was last touched. It hasn't been updated in years and
the content inside it looks outdated.

I would suggest we just delete file and if we need to look into our
translation setup again then we can do so. No point in an old file
laying around if it's not providing value.