feat(context): Build and publish package as dual esm/cjs (#11294)
Followed the general pattern we've established around our approach to
dual publishing. I only wrote one additional package.json file in the
dist because it felt a little redundant to repeat the one we have at the
root.

We've spoke before about perhaps not wanting to ship two sets of type
definitions - one for cjs and one for esm. I know that they're
essentially the same and for a package like this it almost doubles the
published size. For now let's just ship types for both. In the future
we'll stop publishing cjs anyway.

Since I was just following our established pattern I'm going to go ahead
and merge this in. If we discover something we'd rather change we can
always revert or follow up.