fix(auth-providers): Move away from babel for building 'setup' packages  (#11303)
This switches from babel to esbuild for building the 'api' auth provider
packages. Continuing the work of #11301 but this time for the setup
packages.

I have not included the dbAuth package deliberately and will follow up
with a separate update to that package.

I had to make a small change to the `@redwoodjs/cli-helpers` package.
This was because tsc was not happy that I was statically importing from
a package marked `type: module` from a package marked as `type:
commonjs` because it would result in requiring an es module. It was not
understanding that the cli-helpers package was dual cjs/esm and that it
will be totally fine to either import or require from it. The solution
(it appears) is to be specific in the exports field of the package.json.
Those edits here made tsc happy to generate types and didn't throw the
error.

I didn't update the styling of the tsconfig when I changed the
cli-helpers package. We should refactor that standalone. I'm really not
at all thrilled about my recent changes which use the `moduleResolution:
node` because that isn't what we should be using but tsc errors out
based on your `module` and `moduleResolution` setting which has been a
pain. This for now is okay and we can revisit once we encounter the
issue of using two different module resolution algorithms depending on
cjs vs esm.