fix(auth-providers): Move away from babel for building 'api' packages (#11301)
This switches from babel to esbuild for building the 'api' auth provider
packages.

I added `exports` to the package.json. I added all existing files to the
exports field.

I updated the tsconfig files to use `Node16` module and module
resolution. I explicitly set the tsbuildinfo file too. My thinking is
that we can be explicit to start with. Once we have all our packages
over to esbuild and know the lay of the land then we can
deduplicate/refactor our configs.

This is a chore but does effect what gets shipped so I'll label it as
fix and mark it as next-release.