fix(codemods): Move away from babel for building package (#11306)
Switch from babel to esbuild for building the codemods package.

I consider this package a little more fragile than most of ours simply
because it contains code which was written a long time ago. I haven't
removed any dependencies or introduce the exports field here. Can follow
up with that when/if needed.