chore(ci): Temporary fix for format checking on all PRs (#11324)
We currently only run `build-lint-test` if it's a code change but I have
updated the repo so that non-code content - like docs - have to be
formatted correctly too.

This is a temporary fix to allow work to continue and prevent
unformatted content from merging in. I will follow up with a proper fix
that decouples formatting/linting from the building and testing checks.