fix(record): Update dependencies and build dual esm/cjs (#11300)
The main motivation here was to switch the build off of babel to
esbuild. In doing so I've made it dual esm/cjs.

I also noticed that some of the dependencies/scripts didn't look quite
right. I have limited understanding of the details of this package so
I'll work with @cannikin to confirm.