refactor(prerender): build with esbuild and introduce conditional exports (#11337)
Switches the prerender package to be built with esbuild rather than
babel.

This change doesn't not use module or moduleResolution Node16 in the
build version of the tsconfig. This is because it would pick up cjs/esm
issues with apollo and vite. I will look to follow up with fixes that
can enable us to use the modern settings in the build version also.