chore(test-fixture): correct rebuild script string suffix substitution  (#11339)
We replace the randomly generated suffix in scenarios when we rebuild
the test fixture to prevent git conflicts. The regex wasn't quite right.
The number is randomly generated between 0 and 10 million but the regex
required at least 6 digits. There was therefore cases where the
generated number was less than 6 digits and did not get replaced.