fix(jobs): Include bin proxies and only build to cjs (#11344)
The jobs bins where not being picked up when you ran jobs from within a
user project. I have followed our existing approach of adding bin
proxies to the core package. I think we should look to revisit this
general approach in the future but not right now.

I also ran into issues with cjs/esm compatibility for the bins. We need
to build them to cjs as they end up required by another cjs file. For
now I have switched the jobs package to build only to cjs and we can
switch to hybrid or esm only in the future. I would prefer to get this
fix in now so that the feature is fixed and then we can enhance how we
build things.