chore(jobs): Add e2e test script and CI job (#11348)
Adds a test script which will do an end-to-end style test of the redwood
jobs feature. It takes a test project, sets up up jobs, generates a job,
schedules a job, and executes the job.

It can be run locally with just one command. It has been added to CI
too.

---------

Co-authored-by: Rob Cameron <cannikin@fastmail.com>