chore(ci): Pin action dependencies by digest (#11395)
It is considered a best practice to pin github actions by the actual
digest rather than by tag - given that tags are mutable.

There's a convenient tool
[pin-github-action](https://github.com/mheap/pin-github-action) that has
a CLI that can take a file and do this for us. I used this here and
ensured that I had the comments in the form that renovate will
understand. Renovate will trigger updates for the digest when a new
version (based on the comment) is published
[(docs)](https://docs.renovatebot.com/modules/manager/github-actions/#additional-information).

This might make working with actions slightly harder than it was before
but we often use the same actions repeatedly so in that case just copy
paste the digest from somewhere else. If you're adding a new one use the
version tag and then run that tool mentioned above.