chore(publishing): Fix canary publishing for next branch (#11429)
You'll see that the canary publishing for the next branch is broken.
Example action log:
https://github.com/redwoodjs/redwood/actions/runs/10706015686/job/29682729719

The problem as I see it is that we're grep'ing for `-canary.` but in
this case we're looking to publish with the `-next.` prefix which is
clearly the one output by lerna. This PR updates the grep to use a regex
for either the next or canary match.