fix(linting): Re-enable babel presets during linting of javascript projects (#11458)
Fixes #11457 

In #10867 I removed the presets by default and they are only included in
specific cases - for jest, for prerender. It looks like I forgot to
consider the case of linting javascript projects. I therefore re-enable
these presets in that specific case.

It's already on my medium term wish list to simplify and rip out as much
of our babel config/reliance as we can. It would be better if this was
all just simpler and didn't have as many branches.

I also introduced a copy of `isTypeScriptProject` into
`@redwoodjs/project-config`. It doesn't add any dependencies to that
package and it relies on the functionality already provided in it. I
will make a note to refactor our existing usage to use this one. I think
that would be better.