fix(auth0): avoid use of undefined global in browser environment (#11531)
As part of migrating the build tool we likely removed some
transpiliation that was allowing `global` in the browser. Switching to
`globalThis` is a small change which fixes the issue.

I have tested this locally and confirmed the bug before and the fix
after.

We could also directly use `window` instead but I don't really think
there's anything more or less correct about that vs `globalThis`. I
didn't really want to go down the road to adding configuration to
esbuild to start adding back build time changes to support this.