fix(cli): handle both single and double quotes during dbAuth setup (#11534)
Fixes #11526. 

We had a few find and replaces which don't work if you elect to have
double quotes rather than single quotes.

By default we ship with a config that expects single quotes so I am not
going to audit the entire cli/framework in this PR to ensure we support
double quotes at all times. This addresses the issue and so I think is
good enough for now. A lot of these functions are utilities so editing
them here should have good coverage in other commands.