Use node's crypto module for random 6-digit token generation
This is the last remaining place in the guide where 3rd-party crypto-js package is used. Replaced with crypto-js with node's crypto module to simplify the code and get rid of external dependency.

One thing I'm not certain about is how proposed changes compare with the existing solution from the security point of view.  `crypto.randomInt()` should be secure, but after generating a number I padStart the value with "0" which may reduce randomness I guess. However the same is done for crypto-js generated values as well, so I would assume it shouldn't be less secure.