fix(graphql): Allow including 'File' scalar by default to be disabled (#11540)
There was a problem introduced in v8 when we included the `File` scalar
by default. This meant a custom implementation by the user could be
clobbered by the new default. This change allows the user to supply
config to disable including it by default.

This is not how I would have loved to have done things here. Config in
two places is rubbish but given the organisation of this currently it
was generally unavoidable.