Upgrade Prisma from 5.20.0 to 6.2.0
See https://www.prisma.io/docs/orm/more/upgrade-guides/upgrading-versions/upgrading-to-prisma-6

Especially 6.2 brings exciting new features, most importantly `json`
and `enum` fields for SQLite, but also the `omit`-feature (the
antagonist to `include`), the `updateManyAndReturn` batch function,
or the support for auto-generated cuid2 and ulid values.

For a quick overview of the new features, refer to the release notes:
- https://github.com/prisma/prisma/releases/tag/6.2.0
- https://github.com/prisma/prisma/releases/tag/6.1.0
- https://github.com/prisma/prisma/releases/tag/6.0.0

This is, of course, a candidate for the next major version of
RedwoodJS. The migration process looks smooth, we should never-
theless start on a writeup to highlight the new features and
the most important migration steps.