chore(dataMigrate): Switch from jest to vitest (#11978)
As a step towards ESM support we're switching from Jest to Vitest.
This PR takes care of switching our `@redwoodjs/cli-data-migrate` over.

One technique I used here is to alias an import of a CJS package to the
package's source directory, which is using ESM syntax (no `require()`)
so it works with Vitest. (The package's ESM source is then converted to
CJS with `require()` instead of `import` when building)