docs(dbauth): Update the passwordless/OTP how-to (#11994)
I was working through the passwordless how-to and noticed that it hadn't
been updated in a while to match newer RW templates/conventions. So I
updated it as I followed along. It still needs more work. Most pressing
issue is converting it to TypeScript. But I don't have time for that
now, and this is still a big improvement over what we currently have, so
putting this up as a PR.